% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing_parameters.R
\name{add_forcing_attributes}
\alias{add_forcing_attributes}
\title{Generate Forcing Parameters}
\usage{
add_forcing_attributes(gpkg, export_gpkg, add_grid = NULL, dem, verbose = TRUE)
}
\arguments{
\item{gpkg}{hydrofabric geopackage}

\item{dem}{digital elevation model (DEM) SpatRaster to process. Units should be meters}

\item{verbose}{emit messages?}
}
\value{
data.table
}
\description{
The NextGen forcing workstream team has been working on regridding the GFS meteorological forcings data into NextGen catchments.
There is a preliminary step that we must take to “downscale” the regridded data into the NextGen catchments based on previous NCAR formulas
that were integrated into the WRFHydro Forcings Engine for the National Water Model (NWM). The following characteristics are needed to downscale GFS data for each NextGen catchment:
Mean elevation (meters); Catchment Latitude centroid; Catchment Longitude Centroid; Mean Catchment Slope (meters/kilometer); Circular mean azimuth (degrees);
Catchment area (km2)
}
