% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_grid_mapping}
\alias{add_grid_mapping}
\title{Add grid mapping to geopackage}
\usage{
add_grid_mapping(
  gpkg = NULL,
  layer_name = "aggregate_divides",
  template = "/Users/mjohnson/Downloads/AORC-OWP_2012063021z.nc4",
  grid_name = NULL,
  add_to_gpkg = TRUE,
  log = TRUE
)
}
\arguments{
\item{gpkg}{a geopackage file path}

\item{layer_name}{gpkg layer name with polygonal geometries}

\item{template}{a filepath or SpatRast object}

\item{grid_name}{layer name of exported layer when add_to_gpkg = TRUE}

\item{add_to_gpkg}{should the weight grid be added to the gpkg?}

\item{log}{should a log be kept}
}
\value{
file.path or data.frame
}
\description{
Add grid mapping to geopackage
}
