% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_attributes.R
\name{add_lake_attributes}
\alias{add_lake_attributes}
\title{Add Lake Parameters}
\usage{
add_lake_attributes(
  gpkg,
  lake_vars = c("Dam_Length", "ifd", "LkArea", "LkMxE", "OrificeA", "OrificeC",
    "OrificeE", "time", "WeirC", "WeirE", "WeirL"),
  lake_path = NULL
)
}
\arguments{
\item{gpkg}{a hydrofabric gpkg}

\item{lake_vars}{a set of lake variables to extract for lake_path}

\item{lake_path}{a NetCDF path to an NWM LAKEPARAM file}
}
\value{
gpkg filepath
}
\description{
Add Lake Parameters
}
