% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing_parameters.R
\name{forcing_parameters}
\alias{forcing_parameters}
\title{Generate Forcing Parameters}
\usage{
forcing_parameters(gpkg, dem, add_to_gpkg = TRUE)
}
\arguments{
\item{gpkg}{hydrofabric geopackage}

\item{dem}{digital elevation model (DEM) SpatRaster to process. Units should be meters}

\item{add_to_gpkg}{should table be written to the input gpkg (layer_name = forcing_metadata)}
}
\value{
data.table
}
\description{
The NextGen forcings workstream team has been working on regridding the GFS meteorological forcings data into NextGen catchments.
There is a preliminary step that we must take to “downscale” the regridded data into the NextGen catchments based on previous NCAR formulas
that were integrated into the WRFHydro Forcings Engine for the National Water Model (NWM). The following characteristics are needed to downscale GFS data for each NextGen catchment:
Mean elevation (meters); Catchment Latitude centroid; Catchment Longitude Centroid; Mean Catchment Slope (meters/kilometer); Circular mean azimuth (degrees);
Catchment area (km2)
}
