% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwm_grids.R, R/routing.R
\name{latest_nwm_version}
\alias{latest_nwm_version}
\title{Find Latest Version of NWM on NCEP}
\usage{
latest_nwm_version()

latest_nwm_version()
}
\value{
character

character
}
\description{
Find Latest Version of NWM on NCEP

Find Latest Version of NWM on NCEP
}
