% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{ngen_data_dir}
\alias{ngen_data_dir}
\title{hyAggregate data direcory}
\usage{
ngen_data_dir(dir = NULL)
}
\arguments{
\item{dir}{if not supplies will default to \code{get("ngen_dat_dir", envir = hyAggregate_env)}}
}
\value{
character
}
\description{
hyAggregate data direcory
}
