% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_feature_functions.R
\name{realign_topology}
\alias{realign_topology}
\title{Realign Topology to a nexus network}
\usage{
realign_topology(
  network_list,
  nexus_prefix = NULL,
  terminal_nexus_prefix = NULL,
  coastal_nexus_prefix = NULL,
  internal_nexus_prefix = NULL,
  catchment_prefix = NULL,
  waterbody_prefix = NULL
)
}
\arguments{
\item{network_list}{list containing flowpath and catchment \code{sf} objects}

\item{nexus_prefix}{character prefix for nexus IDs}

\item{terminal_nexus_prefix}{character prefix for terminal nexus IDs}

\item{coastal_nexus_prefix}{character prefix for coastal nexus IDs}

\item{internal_nexus_prefix}{character prefix for internal nexus IDs}

\item{catchment_prefix}{character prefix for catchment IDs}

\item{waterbody_prefix}{character prefix for catchment IDs}
}
\value{
list
}
\description{
Realign Topology to a nexus network
}
