% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_download_upload.R
\name{upload_to_sb}
\alias{upload_to_sb}
\title{Authenticated Upload to ScienceBase}
\usage{
upload_to_sb(x, item = "629a4246d34ec53d276f446d")
}
\arguments{
\item{x}{A string vector of paths to files to be uploaded}

\item{item}{a character ScienceBase ID corresponding to the item (default = '629a4246d34ec53d276f446d')}
}
\description{
Authenticated Upload to ScienceBase
}
