% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{add_flowpath_attributes}
\alias{add_flowpath_attributes}
\title{Length Average Routelink Variables}
\usage{
add_flowpath_attributes(
  gpkg,
  rl_vars = c("hf_id", "rl_Qi_m3s", "rl_MusK", "rl_MusX", "rl_n", "rl_So", "rl_ChSlp",
    "rl_BtmWdth_m", "rl_Kchan_mmhr", "rl_nCC", "rl_TopWdthCC_m", "rl_TopWdth_m"),
  hf_version = "2.2",
  source = "s3://lynker-spatial/hydrofabric",
  add_to_gpkg = TRUE
)
}
\arguments{
\item{rl_vars}{routelink variables}

\item{flowpaths}{sf LINESTRING}

\item{rl_path}{routelink path (see get_routelink_path())}
}
\value{
data.frame
}
\description{
Length Average Routelink Variables
}
