% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_feature_functions.R
\name{apply_nexus_topology}
\alias{apply_nexus_topology}
\title{Apply Nexus Topology
This function enforces the nexus-->flowpath topology and adds nexus locations,
a catchment edge list, a flowpath edge list, and a lookup_table to the
network_list object.}
\usage{
apply_nexus_topology(
  gpkg,
  nexus_prefix = "nex-",
  terminal_nexus_prefix = "tnx-",
  coastal_nexus_prefix = "cnx-",
  internal_nexus_prefix = "inx-",
  catchment_prefix = "cat-",
  waterbody_prefix = "wb-",
  verbose = TRUE
)
}
\arguments{
\item{nexus_prefix}{character prefix for nexus IDs}

\item{terminal_nexus_prefix}{character prefix for terminal nexus IDs}

\item{coastal_nexus_prefix}{character prefix for coastal nexus IDs}

\item{internal_nexus_prefix}{character prefix for internal nexus IDs}

\item{catchment_prefix}{character prefix for catchment IDs}

\item{waterbody_prefix}{character prefix for catchment IDs}

\item{network_list}{list containing flowpath and catchment \code{sf} objects}
}
\value{
list
}
\description{
Apply Nexus Topology
This function enforces the nexus-->flowpath topology and adds nexus locations,
a catchment edge list, a flowpath edge list, and a lookup_table to the
network_list object.
}
