% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_feature_functions.R
\name{national_merge}
\alias{national_merge}
\title{Merge VPU NextGen files into single CONUS file}
\usage{
national_merge(
  gpkg,
  outfile,
  layers = c("hydrolocations", "nexus", "flowpaths", "divides", "network",
    "flowpath_attributes"),
  verbose = TRUE
)
}
\arguments{
\item{gpkg}{a set of geopackage paths}

\item{outfile}{a file.path to write to}

\item{verbose}{emit messaging?}
}
\value{
outfile path
}
\description{
Merge VPU NextGen files into single CONUS file
}
